requirements.txt
yfinance
fredapi
pandas
pyarrow
sqlglot
duckdb
pyiceberg[sql-sqlite,pyarrow]
anthropic
python-dotenv
